pyproject.toml
[project]
name = "latex2clip"
version = "0.1.0"
description = "Convert LaTeX formulas in clipboard to images for email"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["latex", "clipboard", "email", "math"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Operating System :: MacOS",
    "Topic :: Text Processing :: Markup :: LaTeX",
    "Topic :: Communications :: Email",
]

dependencies = [
    "pyobjc-framework-Cocoa>=9.0",
    "pyobjc-framework-Quartz>=9.0",
    "matplotlib>=3.7",
    "Pillow>=9.0",
    "rumps>=0.4",
    "watchdog>=3.0",
    "click>=8.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "mypy"]

[project.scripts]
latex2clip = "latex2clip.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
